fix: harden order security and deployment flows
Repair order state transitions, inventory locking, refund idempotency, DTO/PII boundaries, session/SSE behavior, Flyway initialization, deployment safety, and frontend auth/payment handling.

Include focused backend/frontend tests and CI lockfile updates. Exclude generated output, virtualenv files, rendered documents, archives, screenshots, and local scratch files from the rewritten business commit.

+++ backend/src/main/java/com/tibet/tourism/modules/ai/application/AiRouteGenerationJobService.java
@@ -248,7 +248,7 @@ public class AiRouteGenerationJobService {
 
     private RouteJob requireOwnedJob(String jobId, User currentUser) {
         RouteJob job = jobs.get(jobId);
-        if (job == null || currentUser == null || currentUser.getId() == null || job.userId != currentUser.getId()) {
+        if (job == null || currentUser == null || currentUser.getId() == null || job.userId != currentUser.getId().longValue()) {
             throw new IllegalArgumentException("AI route job not found");
         }
         return job;
